requirements.txt
# Web框架
fastapi==0.116.0
uvicorn[standard]==0.35.0
jinja2==3.1.6
python-multipart==0.0.20

# AI和LangChain相关
langgraph==0.6.7
langchain==0.3.26
langchain-openai==0.3.27
langchain-core==0.3.75
langchain-community==0.3.27
langmem==0.0.29

# RAG相关依赖
langchain-huggingface==0.3.1
sentence-transformers==5.1.0
faiss-cpu==1.12.0
# 如果使用GPU，请改用对应的 faiss-gpu 版本

# 文档处理
pypdf==6.0.0
python-pptx==1.0.2
markdownify==1.2.0

# 排序和检索
rank-bm25==0.2.2

# 数据库和缓存
redis==6.4.0

# 工具库
pydantic==2.11.7
tqdm==4.67.1

# 配置文件解析
pyyaml==6.0.2

# 其他依赖
python-dotenv==1.1.1
